+++ src/Umbraco.Web.UI.Client/src/packages/documents/documents/entity-actions/manifests.ts
@@ -1,9 +1,6 @@
 import { UMB_DOCUMENT_DETAIL_REPOSITORY_ALIAS, UMB_DOCUMENT_ITEM_REPOSITORY_ALIAS } from '../repository/index.js';
 import { UMB_DOCUMENT_ENTITY_TYPE } from '../entity.js';
-import {
-	UMB_USER_PERMISSION_DOCUMENT_DELETE,
-	UMB_USER_PERMISSION_DOCUMENT_UNPUBLISH,
-} from '../user-permissions/constants.js';
+import { UMB_USER_PERMISSION_DOCUMENT_DELETE } from '../user-permissions/constants.js';
 import { manifests as createBlueprintManifests } from './create-blueprint/manifests.js';
 import { manifests as createManifests } from './create/manifests.js';
 import { manifests as cultureAndHostnamesManifests } from './culture-and-hostnames/manifests.js';
@@ -12,11 +9,7 @@ import { manifests as moveManifests } from './move-to/manifests.js';
 import { manifests as publicAccessManifests } from './public-access/manifests.js';
 import { manifests as sortChildrenOfManifests } from './sort-children-of/manifests.js';
 import { manifests as notificationManifests } from './notifications/manifests.js';
-
-import {
-	UMB_ENTITY_IS_NOT_TRASHED_CONDITION_ALIAS,
-	UMB_ENTITY_IS_TRASHED_CONDITION_ALIAS,
-} from '@umbraco-cms/backoffice/recycle-bin';
+import { UMB_ENTITY_IS_TRASHED_CONDITION_ALIAS } from '@umbraco-cms/backoffice/recycle-bin';
 
 const entityActions: Array<UmbExtensionManifest> = [
 	{
@@ -36,29 +29,6 @@ const entityActions: Array<UmbExtensionManifest> = [
 			},
 			{
 				alias: UMB_ENTITY_IS_TRASHED_CONDITION_ALIAS,
-			},
-		],
-	},
-	{
-		type: 'entityAction',
-		kind: 'default',
-		alias: 'Umb.EntityAction.Document.Unpublish',
-		name: 'Unpublish Document Entity Action',
-		weight: 500,
-		api: () => import('./unpublish.action.js'),
-		forEntityTypes: [UMB_DOCUMENT_ENTITY_TYPE],
-		meta: {
-			icon: 'icon-globe',
-			label: '#actions_unpublish',
-			additionalOptions: true,
-		},
-		conditions: [
-			{
-				alias: 'Umb.Condition.UserPermission.Document',
-				allOf: [UMB_USER_PERMISSION_DOCUMENT_UNPUBLISH],
-			},
-			{
-				alias: UMB_ENTITY_IS_NOT_TRASHED_CONDITION_ALIAS,
 			},
 		],
 	},
